% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{run_import_data}
\alias{run_import_data}
\title{A function to import data from TGCA dataset}
\usage{
run_import_data(
  df = NULL,
  sheet_numb = NULL,
  headername = NULL,
  tumor_value1 = NULL,
  tumor_value2 = NULL
)
}
\arguments{
\item{df}{dataframe with individuals in rows and cancer attributes in columns}

\item{sheet_numb}{index of the sheet from excel file}

\item{headername}{header name of column we want to filter (no NA present)}

\item{tumor_value1}{Known tumor value1}

\item{tumor_value2}{Known tumor value2}
}
\value{
dataframe corresponding of the sheet number selected and individuals with known tumor status
}
\description{
Apply batch effect correction using function ComBat
}
\examples{

data_tgca_tumor_stat <- run_import_data(df_name_and_path,1,"tumor_status","TUMOR FREE","WITH TUMOR")
}
