% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test.R
\name{stat_test}
\alias{stat_test}
\title{A function to compute a unpaired two-samples wilcoxon test for a TGCA dataset}
\usage{
stat_test(
  df = NULL,
  cancer_type = NULL,
  patients = NULL,
  metadata_col_num = NULL,
  metadata_col = NULL,
  metadata_levels = c(NULL, NULL)
)
}
\arguments{
\item{df}{dataframe with individuals in rows and cancer attributes in columns}

\item{cancer_type}{Actual types of cancer to be selected from (i.e., "ACC", "BRCA", etc...)}

\item{patients}{Actual patient's ID}

\item{metadata_col_num}{headername of numerical metadata colum (i.e., age, os.time)}

\item{metadata_col}{headername of categorical metadata column (i.e., gender, race, vital_status )}

\item{metadata_levels}{two values from the categorical metadata column to be compared (i.e., WHITE, ASIAN for race, or Stage1, stage4 for tumor stage)}

\item{cancer_colname}{header name matching patient's cancer status}

\item{patient_colname}{header name of matching patients' ID}
}
\value{
dataframe corresponding of the sheet number and filtered based on header
}
\description{
A function to compute a unpaired two-samples wilcoxon test for a TGCA dataset
}
\examples{

patho_age <- stat_test(data_tgca_tumor_stat, cancer_type =  c("BRCA") , metadata_col_num= "age_at_initial_pathologic_diagnosis",
                      metadata_col ="ajcc_pathologic_tumor_stage", metadata_levels=c("Stage I","Stage IV") )

gender_age <- stat_test(data_tgca_tumor_stat, patients=data_tgca_tumor_stat$bcr_patient_barcode[grepl("TCGA*",data_tgca_tumor_stat$bcr_patient_barcode)], metadata_col_num= "age_at_initial_pathologic_diagnosis",
                       metadata_col ="gender", metadata_levels=c("FEMALE","MALE") )

vita_ostime <- stat_test(data_tgca_tumor_stat, cancer_type = "BRCA", metadata_col_num= "OS.time",
                        metadata_col ="vital_status", metadata_levels=c("Dead","Alive") )


}
